Menu toggle is back to insert Pressing end while tabbed into the overlay will close it Separated build configurations based on type (basic, driver, and DMA) Changed all uint64_t to uintptr_t to better represent what we are looking for

+++ BaseObject.hpp
@@ -3,7 +3,7 @@
 #include <numbers>
 
 struct BaseObject {
-	uint64_t previousObjectLink;
-	uint64_t nextObjectLink;
-	uint64_t object;
+	uintptr_t previousObjectLink;
+	uintptr_t nextObjectLink;
+	uintptr_t object;
 };
