+++ EFTCamera.hpp
@@ -9,12 +9,12 @@ public:
 	EFTCamera() : EFTCamera(0) {
 	}
 
-	EFTCamera(uint64_t address) {
+	EFTCamera(uintptr_t address) {
 		this->address = address;
 	}
 
 	Matrix GetMatrix() const;
 	Vector3 WorldToScreen(Vector3 point) const;
 
-	uint64_t address;
+	uintptr_t address;
 };
