+++ ListNode.hpp
@@ -3,8 +3,8 @@
 #include <numeric>
 
 struct ListNode {
-	uint64_t first;
-	uint64_t last;
+	uintptr_t first;
+	uintptr_t last;
 
-	uint64_t GetObject(const char* objectName);
+	uintptr_t GetObject(const char* objectName);
 };
