+++ Slot.cpp
@@ -4,7 +4,7 @@
 
 Item Slot::GetContainedItem() {
 	if (!this->cachedContainedItem.has_value()) {
-		this->cachedContainedItem = Item{Memory::ReadValue<uint64_t>(Global::pMemoryInterface, this->address + 0x38)};
+		this->cachedContainedItem = Item{Memory::ReadValue<uintptr_t>(Global::pMemoryInterface, this->address + 0x38)};
 	}
 
 	return this->cachedContainedItem.value();
